api/requirements.txt
python-dotenv
fastapi
uvicorn
requests
jinja2
bs4
lxml
#tensorflow
#transformers